airflow-breweries/requirements.txt
# General packages
pytest==7.2.0
unidecode==1.3.6
python-dotenv==1.0.0
pyspark==3.4.0
httplib2==0.20.2
idna==3.3
protobuf==3.19.3
pyparsing==3.0.6
requests==2.27.1
urllib3==1.26.8
google-cloud-storage==2.8.0